% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmod.R
\docType{data}
\name{riskmod}
\alias{riskmod}
\title{Example Risk Model to Validate for 'rmodcal' Package}
\format{
\code{survival::coxph} object.
}
\usage{
riskmod
}
\description{
Pure cumulative risk model to validate with Cox proportional hazard. Example data \code{\link{data_valid}} can be used to validate this risk model (see example in \code{\link{model.calibration}}).
}
\keyword{datasets}
