% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.calibration.R
\name{var.Eadj}
\alias{var.Eadj}
\title{Variance estimation of adjusted-weighted estimates of the expected number of cases}
\usage{
var.Eadj(risk, auxinfo, ind.ph2, event, incl.prob, gcal, omega = NULL)
}
\description{
Internal use only.
}
\keyword{ModelCal}
