% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.calibration.R
\name{wcov.ncc}
\alias{wcov.ncc}
\title{Covariance between nested case-control samples}
\usage{
wcov.ncc(
  survtime,
  samplestat,
  m,
  psample,
  left.time = 0,
  match.var = 0,
  match.int = 0
)
}
\description{
Internal use only. Source from modelbasedVar() in 'multipleNCC' package.
}
\keyword{ModelCal}
